// src/utils/types.ts
/**
 * Available color names in the theme.
 */
export type AvailableColor =
  | 'primary'
  | 'secondary'
  | 'success'
  | 'warn'
  | 'danger'
  | 'info'
  | ColorPalette
  | string;

/**
 * Available color names in the theme.
 */
export type ColorNameChoice =
  | 'primaryColor'
  | 'secondaryColor'
  | 'backgroundColor'
  | 'foregroundColor'
  | 'successColor'
  | 'warnColor'
  | 'dangerColor'
  | 'infoColor';